Move HlslSourceRewriter.Diagnostics to shared file
