+++ src/ComputeSharp.SourceGeneration.Hlsl/Diagnostics/DiagnosticDescriptors.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace ComputeSharp.SourceGeneration.Diagnostics;
+
+/// <summary>
+/// A container for all <see cref="DiagnosticDescriptor"/> instances for errors reported by analyzers in this project.
+/// </summary>
+internal static partial class DiagnosticDescriptors
+{
+}
